Create RPG Manager.py
On development. The purpose of this program is to be a 'manager' to a RPG like D&D that i was invited and challenged to write a program to help-me on this task. I will let all the code open to receive feedback and if you want to test it, feel free to do.